requirements.txt
PySocks==1.5.6
beautifulsoup4==4.4.1
cffi==1.6.0
cryptography==1.4
fake-useragent==0.0.8
idna==2.1
lxml==3.6.0
psycopg2==2.6.1
pyOpenSSL==16.0.0
pyasn1==0.1.9
pycparser==2.14
requests==2.10.0
six==1.10.0
wheel==0.24.0
